requirements.txt
huggingface_hub
langchain-experimental
transformers
accelerate
langchain
langgraph
langchain.core
langchain-huggingface
torch
selenium
firecrawl-py
langchain_community
openai
langchain_openai
trl
ray[rllib]
gym
dotenv

